\name{BigCal}
\docType{data}
\alias{BigCal}
\alias{Calibration}
\title{Interpolated Northern hemisphere Intcal04 calibration curve}
\description{
A two column file giving the radiocarbon age and standard error for every calendar year starting at -5 years BP (AD1955)
}
\usage{BigCal}
\format{A two column matrix containing 26006 observations}
\source{Intcal04}
\references{
Reimer, P.J., M.G. Baillie, E. Bard, A. Bayliss, W.W. Beck, C.J. Bertrand,
  P.G. Blackwell, C.E. Buck, G.S. Burr, K.B. Cutler, P.E. Damon, R.L.
  Edwards, R.G. Fairbanks, M. Friedrich, T.P. Guilderson, A.G. Hogg, K.A.
  Hughen, B. Kromer, G. Mccormac, S. Manning, C.B. Ramsey, R.W. Reimer,
  S. Remmele, J.R. Southon, M. Stuiver, S. Talamo, F. Taylor, J.v.d. van der
  Plicht, and C.E. Weyhenmeyer (2005), Intcal04 terrestrial radiocarbon age calibration, 0-26 cal kyr bp.
Radiocarbon, 46(3), 1029-1058.
}
\keyword{datasets}
     
