\name{GlendaloughEventDepths}
\docType{data}
\alias{GlendaloughEventDepths}
\alias{EventDepths}
\title{Supplementary depths for the Glendalough core.}
\description{
A single column file giving the depths in cm at which various events may have occurred throughout the deposition 
history of the core. Note that these need not be the same as the ddepths.
}
\usage{GlendaloughEventDepths}
\format{A single column file with 6 rows}
\references{
Haslett and Parnell (2007, submitted to JRSSC)
}
\keyword{datasets}
     
