\name{GlendaloughEventDepthsAlnus}
\docType{data}
\alias{GlendaloughEventDepthsAlnus}
\alias{EventDepths}
\title{Supplementary depths for the Glendalough core.}
\description{
A two column file giving the depths in cm at which the Alnus rise may have occurred throughout the deposition 
history of the core. Note that these need not be the same as the ddepths.
}
\usage{GlendaloughEventDepthsAlnus}
\format{A single column file with 6 rows}
\references{
Haslett and Parnell (2008), A simple monotone process with application to radiocarbon-dated depth chronologies.
Applied Statistics 57 (5), 1--20.
}
\keyword{datasets}
     
