\name{intcal13}
\alias{intcal13}
\docType{data}
\title{
Nothern hemisphere 2013 calibration curve
}
\description{
Northern hemisphere 2013 calibration curve
}
\usage{data(intcal13)}
\format{
  A data frame with 5141 observations on 5 variables.
}
\details{
For full details and reference see http://www.radiocarbon.org/IntCal13.htm. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
