\name{marine13}
\alias{marine13}
\docType{data}
\title{
Marine 2013 calibration curve
}
\description{
Marine 2013 calibration curve
}
\usage{data(marine13)}
\format{
  A data frame with 4801 observations on 5 variables.
}
\details{
For full details and reference see http://www.radiocarbon.org/IntCal13.htm. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
