\name{newgraphwindow}
\alias{newgraphwindow}
\title{ Opens a new graphics window on a variety of platforms }
\description{
    Opens a new graphics window on a variety of platforms
}
\usage{
newgraphwindow()
}
\details{
  Not intended for use outside \code{\link{Bchronmenu}}
}
\author{ Andrew Parnell }
\keyword{ programming}
