\name{normal}
\alias{normal}
\docType{data}
\title{
Data for dummy calibration of normally distributed ages
}
\description{
Data for dummy calibration of normally distributed ages
}
\usage{data(normal)}
\format{
  A data frame with 2 observations on 3 variables.
}
\details{
This is dummy data so that \code{\link{BchronCalibrate}} can calibrate normally distributed dates.
}
\keyword{datasets}
