% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BchronDensityRunFast.R
\name{plot.BchronDensityRunFast}
\alias{plot.BchronDensityRunFast}
\title{Plot run from \code{BchronDensityFast}}
\usage{
\method{plot}{BchronDensityRunFast}(x, plotDates = TRUE,
  plotSum = FALSE, ...)
}
\arguments{
\item{x}{Output from \code{\link{BchronDensityFast}}}

\item{plotDates}{Whether to include individual age pdfs (default TRUE)}

\item{plotSum}{Whether to include sum of age pdfs (default FALSE)}

\item{...}{Other graphical parameters, see \code{\link{par}}}
}
\description{
Plots output from \code{\link{BchronDensityFast}}
}
\details{
Creates a basic plot of output for a run of \code{\link{BchronDensityFast}}
}
\seealso{
Examples in \code{\link{BchronDensityFast}}, and see \code{\link{BchronDensity}}, for a slower, more accurate version of this function
}
