% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BchronRSLRun.R
\name{plot.BchronRSLRun}
\alias{plot.BchronRSLRun}
\title{Plot output from BchronRSL}
\usage{
\method{plot}{BchronRSLRun}(x, type = c("RSL", "rate", "accel"), ...)
}
\arguments{
\item{x}{An object created by \code{\link{BchronRSL}}}

\item{type}{One of 'RSL', 'rate', or 'accel'. If RSL produces a plot of RSL esimtates from the model. If rate, produces rate estimates. If accel produces acceleration estimates.}

\item{...}{Other arguments to plot, see \code{\link{par}}}
}
\description{
Plot output from the \code{\link{BchronRSL}} function
}
\seealso{
\code{\link{BchronCalibrate}}, \code{\link{Bchronology}}, \code{\link{BchronRSL}}, \code{\link{BchronDensity}}, \code{\link{BchronDensityFast}}
}
