\name{print.Bchron}
\alias{print.Bchron}

\title{
Print method for Bchron objects
}
\description{
Creates nice print output for a Bchron object
}
\usage{
\method{print}{Bchron}(x, ...)
}
\arguments{
  \item{x}{
The output from a run of Bchron; see \code{\link{BchronMCMC}}
}
  \item{\dots}{
Other arguments (not used)
}
}
\author{
Andrew Parnell \email{andrew.parnell@ucd.ie}
}
\seealso{
\code{\link{BchronMCMC}}
}
\examples{
## See http://mathsci.ucd.ie/~parnell_a/
}
\keyword{dplot}
\keyword{programming}
\keyword{models}
\keyword{nonparametric}
\keyword{smooth}
\keyword{package}
