\name{Sluggan}
\alias{Sluggan}
\docType{data}
\title{
Sluggan Moss data
}
\description{
Chronology data for Sluggan Moss data set
}
\usage{data(Sluggan)}
\format{
  A data frame with 31 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{ID of each age}
    \item{\code{ages}}{Age in (14C) years BP}
    \item{\code{ageSds}}{Age standard deviations}
    \item{\code{position}}{Depths in cm}
    \item{\code{thickness}}{Thicknesses in cm}
    \item{\code{calCurves}}{Calibration curve for each age}
  }
}
\details{
This Sluggan Moss data can be downloaded from the European Pollen Database: \url{www.europeanpollendatabase.net}. For usage see \code{\link{Bchronology}} or \code{\link{BchronDensity}}
}
\references{
Smith, A. G., \& Goddard, I. C. (1991). A 12,500 year record of vegetational history at Sluggan Bog, Co. Antrim, N. Ireland (incorporating a pollen zone scheme for the non-specialist). New Phytologist, 118, 167-187.
}
\keyword{datasets}
