\name{summary.Bchron}
\alias{summary.Bchron}
\title{
Summary method for Bchron objects
}
\description{
Creates nice summary output for Bchron objects
}
\usage{
\method{summary}{Bchron}(object, ...)
}
\arguments{
  \item{object}{
The output from a run of Bchron; see \code{\link{BchronMCMC}}
}
  \item{\dots}{
Other arguments (not used)
}
}
\author{
Andrew Parnell \email{andrew.parnell@ucd.ie}
}
\seealso{
\code{\link{BchronMCMC}}
}
\examples{
## See http://mathsci.ucd.ie/~parnell_a/
}
\keyword{dplot}
\keyword{programming}
\keyword{models}
\keyword{nonparametric}
\keyword{smooth}
\keyword{package}
