% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BchronCalibratedDates.R
\name{summary.BchronCalibratedDates}
\alias{summary.BchronCalibratedDates}
\title{Summarise a BchronCalibrate object}
\usage{
\method{summary}{BchronCalibratedDates}(object, prob = 95, ...,
  digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{The output of a run of \code{\link{BchronCalibrate}}}

\item{prob}{A percentage value (between 0 and 100) at which the highest density regions for each age are calculated}

\item{...}{Further arguments (not currently supported)}

\item{digits}{Significant digits to display (not currently supported)}
}
\description{
Produces summary output from a \code{\link{BchronCalibrate}} run, including the highest density regions for the calibrated ages for given probability levels
}
\seealso{
\code{\link{BchronCalibrate}}, \code{\link{Bchronology}}, \code{\link{BchronRSL}}, \code{\link{BchronDensity}}, \code{\link{BchronDensityFast}}
}

