% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BchronDensityRun.R
\name{summary.BchronDensityRun}
\alias{summary.BchronDensityRun}
\title{Summarise a Bchron density object}
\usage{
\method{summary}{BchronDensityRun}(object, prob = 0.95, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{Output from a run of \code{\link{BchronDensity}}}

\item{prob}{Probability for identifying phases}

\item{...}{Other arguments (not currently supported)}

\item{digits}{Number of digits to report values}
}
\description{
Summarise a \code{\link{BchronDensity}} object
}
\seealso{
\code{\link{BchronDensity}}
}

