\name{summary.BchronRSLRun}
\alias{summary.BchronRSLRun}

\title{
Summarise a BchronRSL run
}
\description{
Summarise a \code{\link{BchronRSL}} run
}
\usage{
\method{summary}{BchronRSLRun}(object, ...)
}
\arguments{
  \item{object}{
The output from a run of \code{\link{BchronRSL}}
}
  \item{\dots}{
(optional) Other arugments to functions (not currently implemeted)
}
}
\author{
Andrew Parnell <andrew.parnell@ucd.ie>
}
\seealso{
\code{\link{BchronCalibrate}}, \code{\link{Bchronology}}, \code{\link{BchronRSL}}, \code{\link{BchronDensity}}, \code{\link{BchronDensityFast}} 
}
\examples{
# Examples in \code{\link{BchronRSL}}
}