\name{TestRSLData}
\alias{TestRSLData}
\docType{data}
\title{
Relative sea level data
}
\description{
A set of relative sea level data for use with \code{\link{BchronRSL}}
}
\usage{data(Glendalough)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Depth}}{Depth in cm}
    \item{\code{RSL}}{Relative sea level in m}
    \item{\code{Sigma}}{Standard deviation of RSL measurement}
  }
}
\references{
Andrew C. Parnell and W. Roland Gehrels (2013) 'Using chronological models in late holocene sea level reconstructions from salt marsh sediments' In: I. Shennan, B.P. Horton, and A.J. Long (eds). Handbook of Sea Level Research. Chichester: Wiley
}
\examples{
data(TestChronData)
# See \code{\link{BchronRSL}}
}
\keyword{datasets}
