\name{Bchroncheck}
\alias{Bchroncheck}
\title{ Checking of Bchron data objects }
\description{
    A quick check to make sure that the Bchron data object will run in \code{\link{BchronMCMC}}. 
}
\usage{
Bchroncheck(Bchrondata)
}
\arguments{
  \item{Bchrondata}{ A Bchron data object with fields for the the input and output
 files, the name of the core being used, the details of the calibration curve, and details of the 
 model run required. }
}
\references{
Haslett and Parnell (2008), A simple monotone process with application to radiocarbon-dated depth chronologies.
Applied Statistics 57 (5), 1--20.
}
\seealso{ \code{\link{Bchronloaddata}} }
\keyword{dplot}
\keyword{programming}
\keyword{models}
\keyword{nonparametric}
\keyword{smooth}
\keyword{package}

