\name{Glendaloughddepths}
\docType{data}
\alias{Glendaloughddepths}
\alias{ddepths}
\title{Pollen (or design) depths for Glendalough core}
\description{
A single column file giving the depths at which pollen was counted for the Glendalough core in Co Wicklow, Ireland.
The task of the model is to estimate the calendar age of the core at these depths.
}
\usage{Glendaloughddepths}
\format{A single column file with 150 rows}
\references{
Haslett and Parnell (2007, submitted to JRSSC)
}
\keyword{datasets}
     
