% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal13}
\alias{intcal13}
\title{Northern hemisphere 2013 calibration curve}
\format{
A data frame with 5141 observations on 5 variables.
}
\usage{
data(intcal13)
}
\description{
Northern hemisphere 2013 calibration curve
}
\details{
For full details and reference see http://www.radiocarbon.org/IntCal13.htm. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
