% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal20}
\alias{intcal20}
\title{Northern hemisphere 2020 calibration curve}
\format{
A data frame with 9501 observations on 5 variables.
}
\usage{
data(intcal20)
}
\description{
Northern hemisphere 2020 calibration curve
}
\details{
For full details and reference see http://intcal.org/blurb.html. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
