% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine13}
\alias{marine13}
\title{Marine 2013 calibration curve}
\format{
A data frame with 4801 observations on 5 variables
}
\usage{
data(marine13)
}
\description{
Marine 2013 calibration curve. The first 3 columns are the calibrated age (in years BP), the radiocarbon age (in 14C years BP), and the 1 sigma standard error (also in 14C years BP).
}
\details{
For full details and reference see http://intcal.org/blurb.html. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
