% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine13}
\alias{marine13}
\title{Marine 2013 calibration curve}
\format{A data frame with 4801 observations on 5 variables}
\usage{
marine13
}
\description{
Marine 2013 calibration curve
}
\details{
For full details and reference see \url{http://www.radiocarbon.org/IntCal13.htm}. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
