\name{newgraphwindow}
\alias{newgraphwindow}
\title{ Opens a new graphics window on a variety of platforms }
\description{
    Opens a new graphics window on a variety of platforms
}
\usage{
newgraphwindow(width,height)
}
\arguments{
  \item{width}{
Width of graph in cm
}
  \item{height}{
Height of graph in cm
}
}
\details{
  A neat way of opening a new window of desired width/height on a variety of platforms.
}
\author{ Andrew Parnell <andrew.parnell@ucd.ie> }
\keyword{ programming}

