% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normal}
\alias{normal}
\title{Data for dummy calibration of normally distributed ages}
\format{A data frame with 2 observations on 3 variables.}
\usage{
normal
}
\description{
Data for dummy calibration of normally distributed ages
}
\details{
This is dummy data so that \code{\link{BchronCalibrate}} can calibrate normally distributed dates.
}
\keyword{datasets}
