% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BchronDensityRun.R
\name{plot.BchronDensityRun}
\alias{plot.BchronDensityRun}
\title{Plot output from \code{BchronDensity}}
\usage{
\method{plot}{BchronDensityRun}(x, plotDates = TRUE,
  plotRawSum = FALSE, plotPhase = TRUE, phaseProb = 0.95, ...)
}
\arguments{
\item{x}{Output from \code{\link{BchronDensity}}}

\item{plotDates}{Whether to plot the individual calibrated dates}

\item{plotRawSum}{Whether to plot the raw sum of the probability distributions}

\item{plotPhase}{Whether to plot the phase values}

\item{phaseProb}{The probability value for the phase identification}

\item{...}{Other graphical commands. See \code{\link{par}}}
}
\description{
Plot output from \code{BchronDensity}
}
\seealso{
See  \code{\link{BchronDensity}} for examples, also \code{\link{Bchronology}}, \code{\link{BchronRSL}}, and \code{\link{BchronDensityFast}} for a faster approximate version of this function
}
