\name{plot.BchronologyRun}
\alias{plot.BchronologyRun}
\title{
Plot output from Bchronology
}
\description{
Plots output from a run of \code{\link{Bchronology}}
}
\usage{
\method{plot}{BchronologyRun}(x, \ldots)
}
\arguments{
  \item{x}{
The object created by \code{\link{Bchronology}}
}
  \item{\dots}{
Other graphical parameters as detailed in \code{\link{par}}
}
}
\details{
Creates a simple plot of the chronology output. More detailed plots can be created by manipulating the Bchronology object as required
}
\author{
Andrew Parnell <andrew.parnell@ucd.ie>
}
\seealso{
\code{\link{BchronCalibrate}}, \code{\link{Bchronology}} \code{\link{BchronRSL}}, \code{\link{BchronDensity}}, \code{\link{BchronDensityFast}} 
}
\examples{
# Examples in main function \code{\link{Bchronology}}
}
