% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BchronRSLRun.R
\name{plot.BchronRSLRun}
\alias{plot.BchronRSLRun}
\title{Plot output from BchronRSL}
\usage{
\method{plot}{BchronRSLRun}(x, type = c("RSL", "rate", "accel"), ...)
}
\arguments{
\item{x}{An object created by \code{\link{BchronRSL}}}

\item{type}{One of \code{RSL}, \code{rate}, or \code{accel}. If \code{RSL} produces a plot of RSL estimates from the model. If \code{rate}, produces rate estimates. If \code{accel} produces acceleration estimates.}

\item{...}{Other arguments to plot, see \code{\link{par}}}
}
\description{
Plot output from the \code{\link{BchronRSL}} function
}
\seealso{
\code{\link{BchronCalibrate}}, \code{\link{Bchronology}}, \code{\link{BchronRSL}}, \code{\link{BchronDensity}}, \code{\link{BchronDensityFast}}
}
