% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal13}
\alias{shcal13}
\title{Southern hemisphere 2013 calibration curve}
\format{A data frame with 5141 observations on 5 variables.}
\usage{
shcal13
}
\description{
Southern hemisphere 2013 calibration curve
}
\details{
For full details and reference see url{http://www.radiocarbon.org/IntCal13.htm}. For usage details see \code{\link{BchronCalibrate}}
}
\keyword{datasets}
