% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestRSLData}
\alias{TestRSLData}
\title{Relative sea level data}
\format{A data frame with 24 observations on the following 3 variables:
\describe{
\item{\code{Depth}}{Depth in cm}
\item{\code{RSL}}{Relative sea level in m}
\item{\code{Sigma}}{Standard deviation of RSL measurement}
}}
\source{
Andrew C. Parnell and W. Roland Gehrels (2013) 'Using chronological models in late holocene sea level reconstructions from salt marsh sediments' In: I. Shennan, B.P. Horton, and A.J. Long (eds). Handbook of Sea Level Research. Chichester: Wiley
}
\usage{
data(TestRSLData)
}
\description{
A set of relative sea level data for use with \code{\link{BchronRSL}}
}
\keyword{datasets}
