% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unCalibrate.R
\name{unCalibrate}
\alias{unCalibrate}
\title{Uncalibrate a Radiocarbon date}
\usage{
unCalibrate(calAges, calCurve = "intcal13", type = c("samples", "ages"),
  pathToCalCurves = system.file("data", package = "Bchron"), ...)
}
\arguments{
\item{calCurve}{The calibration curve to use. Must be the same length as \code{calAge}}

\item{pathToCalCurves}{The path to the calibration curve directory. Defaults to the location of the standard calibration curves given in the package.}

\item{calAge}{The calibrated age value (can be a vector)}
}
\value{
A vector of uncalibrated ages
}
\description{
Turns a calibrated age estimate into an uncalibrated age based on a given calibration curve
}
\examples{
unCalibrate(2350)

unCalibrate(calAge = c(2350, 4750, 11440), 
calCurve = c('marine13', 'intcal13', 'shcal13'))
}

